Switch to RecyclerView from ListView/GridView
There are some problems still but the hugely increased scrolling
performance alone was worth it. I need to refactor the code
sometime after I've gathered more knowledge around Recyclerview.

The main motivation to switch to RecyclerView was to have the
capability to get the exact scroll change on scrolling and thus
implement a quick return toolbar just like the one in the
Play Store app.