Go back to ListView for PostsDialog. fixes #69
Why go back from `RecyclerView`? As the by default provided `LayoutManager`s do not support `wrap_content` and thus no easy ability to center the inline posts, it seemed to be the best course of action to simply go back to `ListView` in the meantime. Sadly, that means that `PostsFragment.PostsAdapter` cannot be reused which is the reason for `PostsDialogAdapter` and the 'Don't Repeat Yourself'-method definitions.