Make landscape mode and configuration changes work. see #25 and #67
It's by no means a perfect solution but it should be good enough for typical usage.

Besides, I tried to give the post views a maximum width such that they aren't too wide in landscape mode. But since I decided not to bother with activity and fragment recreation I couldn't just use another layout file. But I also couldn't make a 'maxWidth' work on the postview layout because it's super hard to achieve it.