Add subtle drop shadow to toolbar in image fragment
In order to make the toolbar text/symbols more discernable on bright backgrounds.

The implementation is much more intricate and involved than one might first think... It's not just toolbar.setDropShadow(radius, offsetx, offsety, color) or something similar. Instead I use an ImageView that's behind the toolbar and gets a "screenshot" of the toolbar when the image fragment is started. The color of that screenshot is adapted and it is blurred. Additional challenges are configuration changes and the animation of the toolbar. Both have to be considered else the shadow might be in a wrong place.