Finish watchlist. fixes #28
But there are still things like background refresh (#169) or notifications like in Chanu for the future.

Observation: Implementing the watchlist lead to a relatively huge refactoring need and was much more subtle and complex than anticipated. Even as a programmer one often sees the issues only when standing in front of them.